Added parameterized BuildRelease.msbuild file.
This is the same file currently in use for Hyprlinkr. The goal is to
standardise on a single build file, and maintain differences in
configuration files (Build.config).

One problem with using this standardisation is that in the case of
ZeroToNine, it generates a single NuGet package warning. Before, all NuGet
warnings were suppressed, which wasn't a good solution either, but
currently, warnings are not suppressed, and that results in a single,
expected warning.

FWIW, I've posted a comment on the NuGet discussion thread about NuGet
package warnings: https://nuget.codeplex.com/discussions/273522.