Incremented major version.
Looking at the history of this repository, I haven't changed the API of
this code base for more than two years, yet I still regularly use it
(several times a week, usually). I take that as evidence that it's stable,
so it's time to release version 1.0.0.

This version change is purely cosmetic, as there are no changes compared
to version 0.6.2.